--- src/pages/detail/components/DetailHeader.tsx
@@ -1,1 +1,0 @@
-
